convert VLANGroup to have a scope instead of a location
A VLANGroup exists to enforce uniqueness of a VLAN which can be
constrained by different scopes than just a Location. Convert the
VLANGroup object to have a scope where the relationship can be a
Location as it is today but it can also be a Rack, a Cluster, or a
ClusterGroup.

Fixes #6390. Relates to #3657, #3110, #5050.